Basic logic there, need to add pagination
