+++ internal/datastore/interfaces.go
@@ -2,20 +2,18 @@ package datastore
 
 import (
 	"context"
-	"github.com/RogueAlmond70/code-review-challenge/types"
 	"time"
+
+	"github.com/RogueAlmond70/code-review-challenge/types"
 )
 
 // Though postgres has been chosen for this implementation, we want the flexibility of using whatever database client
 // best fits our needs. The DBClient interface has been designed with this in mind.
 type DBClient interface {
 	GetSingleNote(ctx context.Context, userId string, noteId string) (types.Note, error)
-	GetNotes(ctx context.Context, userId string) ([]types.Note, error)
-	//GetAllNotes(ctx context.Context, userId string) ([]types.Note, error)
-	//GetArchivedNotes(ctx context.Context, userId string) ([]types.Note, error)
-	//GetUnarchivedNotes(ctx context.Context, userId string) ([]types.Note, error)
+	GetNotes(ctx context.Context, userId string, archiveFilter *bool) ([]types.Note, error)
 	CreateNote(ctx context.Context, userId, title, body string) (types.Note, error)
-	UpdateNote(ctx context.Context, userId, noteId string, note types.NoteDto) (types.Note, error)
+	UpdateNote(ctx context.Context, userId, noteId string, note *types.NoteDto) (types.Note, error)
 	DeleteNote(ctx context.Context, userId, noteId string) error
 }
